fix: getTopOperationList is moved under __mocks__
